Separate operations and executor in scheduler_perf directory
